+++ packages/app-next/CHANGELOG.md
@@ -5,11 +5,11 @@
 ### Patch Changes
 
 - Updated dependencies [f090030]
-  - @backstage-community/plugin-jenkins@0.11.1
+  - @qortexone-jenkins/jenkins@0.11.1
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [8ce69ed]
-  - @backstage-community/plugin-jenkins@0.11.0
+  - @qortexone-jenkins/jenkins@0.11.0